requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.3,<5.5
python-json-logger>=2.0.7
